finished `server` tests

+++ server/src/controllers/upload.controller.ts
@@ -4,7 +4,7 @@ import { signInWithEmailAndPassword } from "firebase/auth";
 import { Auth, File, UpdateFile, Upload, createAuthSchema, createFileSchema, createFileUpdateSchema } from "../interfaces/upload.interface";
 import { validarUploadArquivo } from "../utils/validator";
 import { z } from "zod";
-import { auth } from "../utils/constants";
+import { auth, uploadReturnMessage } from "../utils/constants";
 import { FirebaseError } from "firebase/app";
 const dotenv = require("dotenv");
 dotenv.config();
@@ -28,13 +28,18 @@ export const uploadUserImageController = async (req: Request, res: Response, nex
       email: authInfo.email,
       senha: process.env.FIREBASE_SENHA,
     };
+
+    if (authFields.email != process.env.FIREBASE_EMAIL) {
+      throw Error("Dados de autenticação inválidos!");
+    }
+
     await signInWithEmailAndPassword(auth, authFields.email, authFields.senha!);
 
     const storageRef = ref(storage, `users/${upload.fileName}`);
     await uploadBytesResumable(storageRef, upload.buffer!, upload.metadata);
     const image: string = await getDownloadURL(storageRef);
 
-    res.status(201).json({ image: image });
+    res.status(201).json({ image: image, message: uploadReturnMessage.upload });
   } catch (err) {
     const errMessage: string = (err as Error).message ?? "Ocorreu um erro ao tentar fazer o upload da imagem! Por favor, tente novamente mais tarde!";
 
@@ -57,18 +62,23 @@ export const uploadUserImageController = async (req: Request, res: Response, nex
 
 export const removeUserImageController = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const authInfo: Auth = createAuthSchema.parse(req.body);
     const file: File = createFileSchema.parse(req.body);
 
     const storage = getStorage();
     const storageRef = ref(storage, `users/${file.fileName}`);
     const authFields = {
-      email: process.env.FIREBASE_EMAIL,
+      email: authInfo.email,
       senha: process.env.FIREBASE_SENHA,
     };
 
+    if (authFields.email != process.env.FIREBASE_EMAIL) {
+      throw Error("Dados de autenticação inválidos!");
+    }
+
     await signInWithEmailAndPassword(auth, authFields.email!, authFields.senha!);
     await deleteObject(storageRef);
-    res.status(200).json({ message: "Imagem excluída com sucesso!" });
+    res.status(200).json({ message: uploadReturnMessage.delete });
   } catch (err) {
     const errMessage: string = (err as Error).message ?? "Ocorreu um erro ao tentar remover a imagem! Por favor, tente novamente mais tarde!";
 
@@ -101,6 +111,11 @@ export const updateUserImageController = async (req: Request, res: Response, nex
       email: authInfo.email,
       senha: process.env.FIREBASE_SENHA,
     };
+
+    if (authFields.email != process.env.FIREBASE_EMAIL) {
+      throw Error("Dados de autenticação inválidos!");
+    }
+
     await signInWithEmailAndPassword(auth, authFields.email!, authFields.senha!);
     await deleteObject(storageRef);
 
@@ -118,7 +133,7 @@ export const updateUserImageController = async (req: Request, res: Response, nex
     await uploadBytesResumable(newStorageRef, upload.buffer!, upload.metadata);
     const image: string = await getDownloadURL(newStorageRef);
 
-    res.status(200).json({ message: "Imagem atualizada com sucesso!", image: image });
+    res.status(200).json({ message: uploadReturnMessage.update, image: image });
   } catch (err) {
     const errMessage: string = (err as Error).message ?? "Ocorreu um erro ao tentar atualizar a imagem! Por favor, tente novamente mais tarde!";
